+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#
+- Fix Container::attach output when TTY is enabled on container
+
 # 0.12.0
 - Fix some integer fields that could be negative but previously were a usize like `ImageSummary::containers`
 - Fix deserialization of nullable map types like `ImageSummary::labels`