+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `SystemDataUsageOpts` parameter to `Docker::data_usage`
 - `Container::stop` and `Container::restart` now take `ContainerStopOpts` and `ContainerRestartOpts` as paramaters respectively
 - Add `VolumeCreateOptsBuilder::cluster_spec` parameter
+- Add `ExecCreateOptsBuilder::console_size` parameter
 
 # 0.12.0
 - Fix some integer fields that could be negative but previously were a usize like `ImageSummary::containers`