+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-#
+# 0.13.0
 - Fix Container::attach output when TTY is enabled on container
 - Add `Image::build_par` that uses multithreaded compression algorithm for creating context directory archive. This method is behind `par-compression` feature flag.
 - Add `ImageListOptsBuilder::shared_size` parameter
@@ -7,6 +7,8 @@
 - Add `VolumeCreateOptsBuilder::cluster_spec` parameter
 - Add `ExecCreateOptsBuilder::console_size` parameter
 - Add `Volume::update` endpoint
+- Bumped default API version and models to v1.42
+- Add parameters Hostname,Domainname,IpcMode and PidMode to `ContainerCreateOptsBuilder`
 
 # 0.12.0
 - Fix some integer fields that could be negative but previously were a usize like `ImageSummary::containers`